Made changest to javascript and html
In my javascript I removed the game() function as specified in the
project requirements. I also added a function that uses an event
listener to get a string that represents the player selection. The
string is taken from the class of each button on the page.

In my html I changed the classes of the buttons in order to represent
the strings I need for the event listener function in my JS.