Added in multiple html, style and js features
For the Javascript functions were added to dsiplay a popup button once
a player/computer reached a score of 5. Also, I completed the functions
to play the game and all of the DOM manipulation related to it.

I added styling to the popup elements in the html that were also added.